+++ src/routes/article-edit.js
@@ -3,16 +3,12 @@
 const { Router } = require(`express`);
 const articleEditRouter = new Router();
 const axios = require("axios");
-const { returnCurrentDate } = require("../helper");
-const emptyPost = {
-  title: ``,
-  announce: ``,
-  fullText: ``,
-  date: returnCurrentDate(),
-};
+const { returnCurrentDate, upload } = require("../helper");
 const type = `Редактирование публикации`;
 const moment = require("moment");
 const { articleValidator } = require("../express/middlewares/validator");
+const bodyParser = require(`body-parser`);
+const jsonParser = bodyParser.urlencoded({ extended: true });
 
 articleEditRouter.get(`/:articleId`, (req, res) => {
   axios
@@ -43,13 +39,15 @@ articleEditRouter.get(`/:articleId`, (req, res) => {
     });
 });
 
-articleEditRouter.put(`/:articleId`, articleValidator, (req, res) => {
+articleEditRouter.post(`/:articleId`,  jsonParser,
+  upload.single(`avatar`), articleValidator, (req, res) => {
+
+  // console.log('PUT req body', req.body);
   axios
-    .put(`http://localhost:3000/api/articles/${req.params.articleId}`, {
-      timeout: 1000,
-    })
+    .put(`http://localhost:3000/api/articles/${req.params.articleId}`, req.body)
     .then((response) => {
       const data = response.data;
+      // console.log('PUT req body in req to back', req.body);
       const fetchedPost = {
         title: data.title,
         announce: data.announce,
@@ -59,13 +57,18 @@ articleEditRouter.put(`/:articleId`, articleValidator, (req, res) => {
         ),
         category: data.category,
       };
-      res.render(`article-add`, {
-        article: fetchedPost,
-        type,
-        pageTitle: type,
-        category: data.category,
-        action: `/articles/edit/${req.params.articleId}`
-      });
+
+      if (Object.keys(res.locals.errorList).length) {
+        res.render(`article-add`, {
+          article: fetchedPost,
+          type,
+          pageTitle: type,
+          category: data.category,
+          action: `/articles/edit/${req.params.articleId}`
+        });
+      } else {
+        res.redirect(`/my`);
+      }
     })
     .catch((err) => {
