+++ src/service/middlewares/validator.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const ArticleKeys = [`title`, `announce`, `fullText`, `category`];
+const ArticleKeys = [`date`,`title`, `announce`, `fullText`, `category`];
 const { HttpCode } = require(`../../HttpCode`);
 const { sendResponse } = require(`../../utils`);
 
@@ -24,11 +24,12 @@ const articlePutValidator = (req, res, next) => {
   const newArticle = req.body;
   const keys = Object.keys(newArticle);
   const keysExists = ArticleKeys.some((key) => keys.includes(key));
-  console.log("Got article in validator on back", newArticle);
+  // console.log("Got article in validator on back", newArticle);
   if (!keysExists) {
+     console.log("article DOES NOT pass validation on back", Object.keys(newArticle).join());
     sendResponse(res, HttpCode.BAD_REQUEST, `no such fields in offer`);
   } else {
-    console.log("article passes validation on back");
+     console.log("article passes validation on back");
     next();
   }
 };
